requirements.txt
discord.py
httpx
google-auth